Created explosions when enemies are destroyed
I had to make a blue explosion to test this out because the default
coloring system isn't working like I want it to. I dont want it to just
take the B channel of the RGB image: i want it to actually change the
Hue of the image, so white is still white, not blue!