Added player collision detection with enemies
Now the player can collide with enemies. But right now it doesn't kill
the player. Also, I moved some methods into classes to organize the code
better and created a new struct that I plan on using for tracking the
player's status when I do eventually make the player be able to die.