Added enemy bullets
AH YEAH, ladies and gentlemen we have DANMAKU!
You can't see your hitbox right now but it's pretty small, about 15
pixels in diameter.
Go ahead and try to dodge!
P.S. The shoot programming is just for testing, we will probably need to
make a better system for danmaku scripting in the future.